requirement.txt
docx
fastapi
uvicorn
Pillow
opencv-python
pathlib
matplotlib
seaborn
scikit-learn
pandas
flask
reportlab
jupyter
python-multipart
numpy
tensorflow